+++ src/test/java/lotto/domain/BuyerTest.java
@@ -1,7 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
