+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,8 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assert;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
