+++ src/test/java/lotto/domain/QuickPickGeneratorTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
